Javascript still not picking up "BETWEEN" values.
Ive temporarily committed my "search.html" test document.  This points
to the JavaScript and CSS which I am using to render search document.  The
first item in the search form "State", has "BETWEEN" available.  However,
Javascript does not seem to be able to get the extra values, even though
my tests work correctly.  When I try to submit, I have the following 
embedded in my query string:

  state_logical=&_state_comp=BETWEEN&state=&state=3&state=4

As you can see, the "state" values are there.  Why they are not visible to
the Javascript is beyond me right now.  Will investigate further tomorrow.
I have also temporarily include a JS version of Data::Dumper.  It is choking
on most values I pass it.  I do not know if this is becuase it is broken or
if this is symptomatic of what I am trying to pass into it.

I have also removed unused exception messages and I fixed a caching bug
in the search objects which was causing storesearch.t tests to fail.  I
should revisit that, though, as I think my search_method() method should
have smarter caching rather than the other methods knowing when to break
the cache.