Failed tests down from 71 to 13.  Getting there.
* Quotes in searches now work.
* Fixed bug where switching from one "non-BETWEEN" search to another
  "non-BETWEEN" search would corrupt the data in the first text box.
* LIKE searches now work.

The search interface is now completely useable with the exception of
ANY searches.  Well pull that out and perhaps put it back in when its
useful.

For now, the main goal is to get the tests working and then pull XML
generation out of the REST class.  I now have some ideas on a way I can 
approach that without *too* much difficulty (I hope).

After that, its time to start pulling widget metadata from attributes
and using this in conjunction with default value lists to make the search
form more effective.