All tests once again pass.
* Converted many is_xml tests to is_well_formed_xml tests.

* Updated localization tests to also search in test directories.

The latter was an annoying problem.  As it turns out, some of the localization
messages are generated, not hard-coded.  Thus, the tests failed to identify
them.

Also, I, getting a lot of "unitialized value" errors in Class::Meta::Class
line 249.  I looked at the code and was not sure what the Right Thing was.
The \@list which gets generated might be empty and I do not know if this is
acceptable behavior or not.

Also, I spent a few minutes packaging up HOP::Parser and HOP::Lexer to get
them on the CPAN and out of our code base.  Hopefully others will be able to
supply patches and fixes as needed.