Refactoring to pull REST XML generation out of the dispatch class.
Much of the refactoring is done, but there are a few methods which are proving
stubborn.  Until I finish, Im leaving the new Kinetic::XML::REST class 
embedded in the dispatch class.  This allows the tests to continue passing and
makes it much easier to continue with the refactoring.

This should be the last major change necessary before I continue on to using
the Kinetic::Meta widget functionality.