Pulled most of the XML generation out of the dispatch class.
Added more tests for the XML generation.

Greatly improved the validation.

Theres more I can do, but things are just getting fiddly at this point.  Too
much work to refactor this without a clear direction is likely to produce
results which will need to be refactored *again*.  I would rather not do that.

Instead, Ill move on to the widget stuff tomorrow.