First pass at Kinetic::Format::JSON.
JSON is a breeze to work with.  The module is easy to use.  Its almost like
using Data::Dumper but with a JSON format.  So far we can render, restore and
save objects to and from JSON.

There is a slight but in my implementation and thus a failing test.  Il fix
that tomorrow.

I have also removed the _get_class_from_key method in Kinetic::Store::DB.