Chained method calls work.
I have the basics of chained method calls in URIs working.  It is not 
particularly clean or well-designed, but it is not too bad, either.  Right now
I have focused on getting it to work and then I will worry about clean-up
later.  Comments welcome.

Eliminated some unused methods in both the REST and Dispatch classes.  Cleaned
up how we were fetching $class and separated the fetching of results from the
store and the conversion of results with the formatter.

The REST class has a phenomenally ugly _get_request_from_path_info method
which surprised me by running correctly the very first time I tried it.

The dispatch class methods now disambiguate between "handling" methods and
"executing" them.  The former is dealt with internally and the execution is
performed by the data store.

Also added the "bad request" result for the dispatching.