Basic Catalyst interface.
Run catalist/server.pl to start the server.  You should be able to login with
theory/theory or ovid/divo.  If not, it is the same problem we had before.

Trying to autogenerate a model, but Kinetic::Meta->keys is not returning any
keys.  Do not know what I have done wrong so I am not checking anything in with
that.

I realize that I will have to replicate the old "search" interface or something
similar as a CRUD interface does not make much sense if you cannot find what 
you are looking for.