Tiny commit to handle some rendering changes.
Added IDs to all HTML attributes.  Added labels. However, they are in  different
spots and this is bad.  Before I fix this, I need to ensure that I move the
plugin to a reasonable spot and then start work on the view/search/edit
rendering and that will sort most of this out (just trust me on that).