Fleshed out more of the search/edit interface.  Right now you can see that it is close to where it was, though it is not there yet.
It does not yet do searching, but there are a few more tests for it.

I also noticed that it was hard to do some of the tests if I did not have
labels or tips set up for everything (due to localization issues).  Do we want
it to be the case that we always fail if those are not supplied?

Also, I noticed that checkbox widgets can be "checked", but I have no idea what 
that is suppposed to mean if we have a boolean value which defaults to false.
Is this just some artifact copied from documentation?