You can now search for and click on Kinetic objects and "view" them.
Moved TT plugins to the Kinetic::UI::TT::Plugins:: namespace.

Added a bit more documentation.

Eliminated FK constraint errors while running REST tests.

Made the default list of keys on the left side of the Catalyst UI do a squery.

Removed apparent duplicates in search interface.

We can now perform "empty" searches.  That was a bug which crept into the
interface.

Basic search results are now in a table below the search interface.  I am not
bothering to make the search interface persistent because even though I know
how, it takes time to do and it is really not relevant to what we currently
need.

Moved Kinetic::UI::Catalyst::V::TToolkit to Kinetic::UI::Catalyst::V::TT.

Next I will start on the "new" and "edit" interfaces (they will probably be
the same thing).  I also need to figure out the best way to test this and also
to get rid of the annoying "login failed" messages in tests.