Started work on Kinetic::Engine::Request.
I have a mock Kinetic::Engine object which currently does not do much but
provide enough information for the request object to understand how to
dispatch.

Trying to set up the request tests to have a single set of tests for shared
behavior amongst request types and I will be able to test their differing
behavior separately.  I am not sure how easy it is to follow the 
test_shared_behavior() method, though.

Since I am not terribly familiar with the Apache stuff, if you can look over
the basics and let me know if it is sane, I would appreciate that.

Merry Christmas.