Basic kineticd script in place.
I backed out a number of changes that I made because I was not sure of the 
best approach and now kineticd only works for Catalyst, not Apache.  This is
because my Apache installation is in a non-standard place (/Library/Apache2)
and I think it would be better if the script would actually handle that
correctly.

I also backed out my work in making "simple" the default server.  The t/build.t
changes were easy enough, but I was not entirely certain of the best way to
handle the change with t/config.t.  And since we are on the topic, I do not
know if my kineticd setup is even the best way to go :/

%SERVERS now point to the appropriate engine class.

Switched apache2 references back to apache, except for the engine class in
Kinetic::Build.

I think I should do some more work this weekend.  I just did not get enough 
done today.