Pulled out $ENV{KINETIC_ROOT} and set it in the conf file.
Added help and debugging info to kineticd

kineticd now supports start and stop commands for Catalyst, not Apache.  I'll
add "restart" and when that's done, start working on making that happen for
Apache and consolidating the code.  Currently, though, it has some serious
problems.  For example, the KINETIC_CONF env variable is not being passed
through and I can't figure out how to get the catalyst log info to be
redirected to a file.  Instead, it's all getting dumped to STDOUT.  Hopefully
(?) I'll get all of this working tomorrow.

Added logs/ directory.  Currently this is only for pid file management as I
wasn't sure where else to put it.  I don't really like this.

Added bin/kineticd to Build.PL so it can be installed in as an executable

Removed the http_ prefixes from thhe apache section of the conf file

Moved Kinetic::Build::Engine::Apache to Kinetic::Build::Engine::Apache2