Lots of work trying to clean up kineticd.
Localized messages in kineticd.  Some localization messages in en_pm will
likely need to be moved to other sections as needed.

Add location of Apache conf file to kineticd and then pulled it back out after
I couldn't get Apache to load it.  I've read the docs, but they don't seem to
be correct.

Cleared up some documenation regarding how KINETIC_CONF works.

Gave uses the ability to specify the location of the httpd.conf file.

I'm trying to make kineticd work with Apache2 and so far I haven't quite
gotten there.  It's a lot further along then it was, but I've noticed that
Proc::Background seems to be picking up the wrong pid.  I don't know if this
is something I'm misunderstanding or if there is a bug.  Today's been a gring
as a result.