Made kineticd work with Apache.
I finally have kineticd working with Apache.  I had to write a routine which
guesses where the Apache pid file is located.  If it can't be found, kineticd
dies.  I added a switch to allow the user to specify the location of the pid
file, if necessary.

For some reason, I sometimes have to run "kineticd stop" twice to get it to
stop the apache server.  This appears to be because File::Pid sometimes
incorrectly reports false to "$pidfile->running".

Kinetic::Engine now loads the Kinetic classes.  For the time being I've
hard-coded t/sample/lib in it just to make sure it works.  Once we know where
we'll put the real Kinetic classes, I'll be able to change that over.

Had to roll back some of the Kinetic Catalyst code I worked on.  Seems that
bleed Catalyst straight out of SVN fails badly (well, it's bleed, so it's not
surprising that I got cut).  My code at the present time no longer relies on
the Catalyst patch I provided.

Overrode Catalyst::Log::_flush with a custom method to allow us to control
what happens with log data (rather than just dumping it to STDERR).

Had to update some Catalyst tests.  Due the way that Catalyst handles
overriding logs, Kinetic::UI::Catalyst must be loaded prior to the rest of the
tests running.  This was not documented well in Catalyst, but I've committed a
small change to the Catalyst docs for this.

Removed t/request.t since its tests were now failing and we aren't using our
own request object.  Also removed all of the request modules since we're not
using them.

Also, is there supposed to be a t/media_type.t?  I get warnings from Build.PL
about it not existing.