Added memcached support for Catalyst sessions.
I was very pleased that memcached was relatively easy to configure and use.
That made life much easier.

Using Catalyst::Plugin::Session::Store::Memcached is now one of the build
options, along with file based sessions.  The latter are useful when you want
to just test Kinetic and don't have memcached installed.

Because of the way memcached works, I did not have kineticd try to start the
memcached daemon.  I was going to, but since those daemons are frequently (if
not always) running on different boxes, kineticd has no way of reliably
locating and starting them.  Instead, the build process asks which caching
scheme you want to use and if it's memcached, it prompts you for IP addresses
and the port of memcached for each address.

Currently, it defaults to '127.0.0.1:11211' if and only if accept_defaults
returns true.  This is because if the user wants to specify their own
memcached information, they can enter multiple IP addresses and that turned
out to be a bit more complicated than I liked if I allowed defaults.  Instead,
there is *no* default IP address available unless accept_defaults is chosen.
The default port is always 11211.

Since this is being used for sessions in Kinetic::UI::Catalyst, I realized
that constantly referring to it in the code as the "cache" might be confusing,
but I went with that anyway since memcached can be used for much more than
just session management.  I can go through the code and change this if it
turns out to be unworkable.

13 tests are now failing.  Some will be quick fixes, others will not.  I hope
to take care of all of those tomorrow.

I still don't have qpstmpd running :(