About half of the failing tests were fixed.
Removed some tests which checked the apache user since we are no longer asking
for this information.

Fixed the tests to realize that we are also asking for the caching scheme.

Cleaned up some of the docs and added to them.

Lost a lot of time debugging where some errors were being thrown.  I cannot
completely replicate it, but Config::Std (which does not use "strict", I might
add) also does not use "warnings".  As a result, Damian did not seem to notice
how noisy that class can be when run under warnings.  This causes a lot of 
problems when we add data to the configuration hash.  Once I finally nail this
down, I will figure out how to fix it on our end or send him a patch.

As a result of the warnings, I failed to notice some real tests were failing.
I will get to those tomorrow.

Side note:  also spent a lot of time today working with Vibe/Spin on poll and 
JavaScript issues.  I think this qualifies as a support issue even though it
was a bug in our code.