Have a basic lexer in place for email messages.
Basically, I set up a standard token format of 

 PERFORM, ACTION, [STAGE]

Where "STAGE" is optional.  After a bit of tweaking, I managed to get the
lexer to discard data which does not immediately appear to be relevant to the
workflow commands.  This may be a mistake and we may go with the JUNK token
format (not useful for parsing, but useful for logging).

The lexer does a bit of overtime to discard anything which doesn't fit the
above format.  HOP::Parser can't parse anything which doesn't fit a standard
BNF grammar, so I can't easily include the junk stuff, hence the excluding.

Still haven't figured out "notes", but we'll get to that tomorrow.