Basic lexer and parser are done.
Currently, the lexer looks like it's in pretty good shape.

I'm also fairly comfortable with the parser.  Currently, it has a very tiny
grammar and returns a simplistic IR, but I think it's enough for our needs.  I
can actually parse moderately complex emails in the forms we discussed in
email.

The docs aren't complete, but I figure that I can flesh them about better as
we get more use out of this and see what the strengths and weaknesses are.