Stub Workflow email handler in place.
In the etc/ directory I've checked in "check_workflow".  This is the shell of
the basic functionality needed for a qpsmtpd filter to handle workflow email.
Right now it doesn't do much other than demonstrate how to validate whether or
not an email is for workflow and how commands might be handled.

There's also a stub for the UI workflow object.  It doesn't do much, but it
does have the basic logic in place for handling workflow emails.  There are
also stub tests in place for when we can flex this out a bit more.

Cleaned up some of the POD.

Upgraded to Tiger last night and spent some time this morning not realizing
why I couldn't SQLite to compile.  Seems I forgot to install the XCode tools.
Also got Postfix up and running again.

There is no documentation for qpsmtpd.  I did a lot of googling to try to
figure out how to get it to discard Workflow email.  Unfortunately, that
behavior was eliminated from qpsmtpd over two years ago.  So far no response
on the list to queries about it.