Converted qpsmtpd plugin to use a trait, thus allowing for easier testing.
Kinetic::UI::Email::Workflow::Qpsmtpd has been created.  This has the methods
which were previously in the qpsmtpd plugin, but are now flattened into that
plugin via a trait.  However, I was using "DONE" for "DISCARD", per the
suggestions on the qpsmtpd list, but it's causing qpsmtpd to hang.  I'll
figure it out tomorrow.

Kinetic::UI::Email::Workflow has been reworked.  The old constructor is there
(probably not much need for it, though) and a new constructor allows this:

 my $wf = Workflow->new( $from_address, $workflow_address );

It tears apart the workflow address to find out the workflow, job and
organization.  I have some stub error checking that should be converted to an
exception, but I don't know if it's needed (yet).  This is because there are
two ways of creating the WF object and the old one (hashref of properties) may
go away.  If that happens, the exception is superfluous.