Basic caching framework for Kinetic started.
There are notes in the technical spec to the effect a user may choose to not
use caching.  Since sessions require some place to store their data (and this
is typically a cache), Kinetic session objects will be required if caching is
not chosen.  Because such a beast does not yet exist, I've deferred
implementing the "no cache" option.  However, I've separated the configuration
for the Catalyst session cache class and the Kinetic store cache class.  This
will make things easier to implement in the future.

Added in "expires" and "root directory" for file based caching.  I'll probably
want expiration for all caching as items older than a certain time should
probably be routinely flushed from the cache.

Refactored some duplicate code in Kinetic::Build which adds cache and engine
data to the config file, but I think the refactored code is less clear than
the duplicate code.  I might revert it.  (Note:  I just reverted it.  It
caused a conflict on the latest svn update).

The Kinetic::UI::Catalyst::Cache class now grabs Catalyst specific caching
information, thus allowing us to potentially use separate caches for Catalyst
sessions and Kinetic objects.

The Kinetic::Util::Cache API is loosely described in its docs, but the
implementation is not there yet.  I should get this finished tomorrow.  The
implementation is based on the memcached recommendations.

The extra caching files listed in Build.PL are the underlying cache files used
for the Catalyst session plugins.

Also submitted a patch to Test::Class which allows running individual tests.
Of course, that's not in this commit, but it does make testing easier.