Folded separate caching tests back into t/util/cache.t
Basically, we want all caching mechanisms to present the same API.  Thus,
duplicating the tests is counter-productive.  In the future, if separate
caching mechanisms offer different capabilities, we can revisit this decision.

Switched from Clear() to clear() in the file based caching as this respects
namespaces (also eliminates a subtle cache-expiration bug which happens with
the global Clear() -- that was real fun).

Also made the expire time apply to all caches, not just file-based caching.

Finally got the memcached caching to work.  Try as I might, I couldn't get
Cache::Memcached::Managed to work correctly, even though it's extensively
tested.  Unfortunately, this means dropping back to Cache::Memcached.  Now, to
handle this properly, I have to maintain an internal list of IDs which have
been added and removed from the cache since.  Further, since namespace support
is not inherent in Cache::Memcached, I've added a "_make_safe_id" method to
ensure that if something else adds something to the cache, there's no worry
about it clashing.

I've also updated the cache.t test to call "clear" at the end of the test.
This ensures that garbage data is not left over at the end of the test.
However, if caching tests fail in the middle, you might wind up with garbage
data and have to kill and restart memcached.

Backed out initial work trying to implement caching in the data store as the
test suite was slowed so much that it ground to a halt.

Note to self:  need to refactor duplicate expiration handling in
Build::Cache::Memcached and Build::Cache::File.

Changed all cache references of UUID to ID.  Unfortunately, the latter doesn't
make it clear that these IDs should be guaranteed to be distinct for different
objects.  I should note that in the docs.