Converted the Setup:Cache classs to use FSA::Rules. Moved all cache configuration in kinetic.conf under [cache]. Renamed the commands --cache-expires, --cache-root, and --memcached-address. They can be configured via the command-line or via prompts. Multiple --memcached-address options can be specified. Woo! :-)
Kinetic::Build now asks for input for stuff in a sensible order. That is, it asks
for Pg stuff after you have selected the store, for cache stuff after you have
selected the cache, etc. It also eliminates some code when adding new stuff to
check (though it could be even less; more to come).

Other stuff: Fixed a bunch of documentation pastos, moved --cache-expires support
into Setup::Cache (instead of both File and Memcached), fixed the get_reply() to
support the display of an array of options (when specified via the command-line),
and removed deleted files from MANIFEST.