Fixed a lot of niggling issues with caching related stuff.
I never figured out what was going on with the file-caching performance, but
absent a demonstrable problem outside of our test suite, this will be ignored
for now.

Changed CACHE_OBJECT to CACHE_OBJECT_CLASS.

Removed a bogus caching test.

t/util/cache.t now uses a separate directory for the caching tests and removes
the directory when finished.  It also SKIPs the performance tests.  They take
a long time to run, but I might want to revisit this.

Removed _expire_time_in_seconds from the cache system.

This is a small commit as I am feeling very under the weather.  I'm going to
have to make this up later this week along with the half day I missed.  Sorry
about that.