Started working on Kinetic::AppBuild again. The goal was to get it to always do the right thing relative to the install_base and, more specifically, kinetic.conf. It is required to find a kinetic.conf--otherwise there's nothing to install a Kinetic app into!
In the process, I discovered that setting install_base_relpaths for each
element I've added (plus lib) is better than setting the install_paths hash
ref, because those settings are always relative to the install_base. This is
important because Kinetic::AppBuild can *change* the install base, if it's
found an alternate path in the KINETIC_ROOT setting in kinetic.conf.

Other stuff:

* I moved the Kinetic::AppBuild tests directly into t/app_build.t and deleted
  t/app_build. Using Test::Class for this was silly, it turns out.

* I turned on the svn:keywords Id property for new files that were missing it.
  How do I do this, you ask? Like this:

    rgrep -l \$Id | xargs svn propset svn:keywords Id

  Just FYI.

* Updated MANIFEST (again).