Fixed failing JSON tests.
The JSON formatting tests were failing with PostgreSQL tests but not with
SQLite.  After a bit of research, it was determined that this was due to how
DBD::Pg handles boolean values.  The interim solution was to stringify all
values as long as they are defined (turning "undef" into the empty string is
bad).

A long term solution would be to have those values properly format themselves
based upon their context.  For example, in JavaScript, a boolean value should
be 'true' or 'false' (and not quoted in the JSON) rather than a 0 or 1 string.