Kinetic::Contact and Kinetic::Type::Contact now use Class::Meta::Declare.
Also corrected many spelling errors.  In fact, that's the bulk of this commit.
The program I used to find the spelling errors was:

 use Test::More;
 eval "use Test::Spelling 0.11";
 plan skip_all => "Test::Spelling 0.11 required for testing POD" if $@;
 all_pod_files_spelling_ok();

Unfortunately, we have tons of words (Kineticode, TKP, workflow. DSN, DBI, and
so on) which Test::Spelling does not recognize.  We could build up a library
of words to use as "stop words", but I didn't think it was worth the effort.
As a result, I did not check in t/pod_spelling.t.