Merged the trunk into the branch.
Seems my version of DBI was 1.48.  Upgrading to 1.50 made some test failures
go away, but not all.  Still researching.  The failures were along the lines
of:

 Can't locate auto/DBI/installed_d.al in @INC

This commit is only the merge from trunk to the branch, with the exception of
my bumping up the required DBI version number to 1.50.

t/app_build.t and t/build.t are failing.  Looking into that now.