Converted Kinetic::Util::Collection to represent an ordered set.
Internally, this was mostly a matter of switching it over to Array::AsHash
since the keys of a hash are, by definition, a set.

In order to ensure we had unambiguous keys, everything inserted into a
collection must have a uuid() method.  Other than that, anything goes.

Also updated Array::AsHash to support a couple of things I needed.