Eliminated a couple of UNIVERSAL::isa calls.
I didn't do this everywhere, though.  I just did it in a couple of places it
felt really safe to do, though as David pointed out, it should be eliminated
completely.

I also switched the unique constraint on collection tables to only the parent
key and the rank.  Otherwise, we could have duplicate ranks for the same
collection.

Most of today was primarily research.  The last major issue that I know of for
the "has_many" relationships is ensuring type safety.  Unfortunately, I need
to know the approach to that before I can alter Kinetic::Meta::Attribute to
handle this correctly.  Once this is resolved, I can start integrating
"has_many" types into the data store.