Almost done with has_many relationships for building classes.
As soon as I nail the last issue or two here, I can move on to implementing
this for the data store.

I eliminated the bugs with the Collection->splice method by deleting that
method.  It's not entirely clear to me how this method would be used, so it
didn't seem worth the effort to retain it.

The new collection types appear to be getting added correctly.

So far I am getting a test failure when trying to add a collection of the
correct type:

not ok 111 - ... but adding a valid collection of the right type should
succeed

#   Failed test '... but adding a valid collection of the right type should
#   succeed'
#   in t/meta.t at line 380.
# died: Kinetic::Util::Exception::ExternalLib (Can't locate object method
# "_add_modified" via package "MyTestHasMany" at
# lib/Kinetic/Meta/AccessorBuilder.pm line 348.
# 
# [lib/Kinetic/Meta/AccessorBuilder.pm:348]
# [lib/Kinetic/Meta/AccessorBuilder.pm:115]
# [/usr/local/lib/perl5/site_perl/5.8.7/Test/Exception.pm:167]
# [t/meta.t:380]
# )

I don't know what this error is.  It's rather mysterious to me.  How does the
"_add_modified" method get added to the MyTestHasMany class?

 my @thingies = map { MyTestThingy->new } 1 .. 3;
 my $coll = Kinetic::Util::Collection::Thingy->new( {
     iter => Iterator->new( sub { shift @thingies } )
 } );
 my $has_many = MyTestHasMany->new;
 
 throws_ok { $has_many->thingies(1) }
     'Kinetic::Util::Exception::Fatal::Invalid',
     'Adding an invalid type to a has_many relationship should fail';
 
 lives_ok { $has_many->thingies($coll) }
     '... but adding a valid collection of the right type should succeed';

The throws_ok fails as expected.  The lives_ok dies fails with the error
message above.