All classes build with Kinetic::Meta automatically inherit from Kinetic.
Because I ran into a few problems with this, I decided to go ahead and
explicitly remove the 'use base "Kinetic"' that I found in many classes.  This
allowed me to shake out some weird assumptions in the code in a few places,
including classes which weren't inheriting from Kinetic.

For some reason, adding Kinetic to the inheritance tree was not sufficient for
dealing with this issue.  I have to eval the "use base 'Kinetic'" line due to
some weird magic that "base" seems to go through.  That wasn't fun to debug.
I still don't know exactly what caused the issue, but it's resolved and I
don't think the eval is really bad, particularly since it's a one-time hit.

Removed some a duplicate test in t/meta.t.  It was failing and didn't appear
to contribute anything.

Fixed the Template Toolkit rendering tests so they were properly using a
Kinetic class.

I think I can now move on to actually getting the "has_many" relationships
working with stores.  I already know of some issues I have, but I'll tackle
them as they come up.

Oh, and I know that there's a good chance this won't happen, but I really hope
you open-source this :)