Collection attributes for objects work.
They need to be redone, though.  I have a rough hack in place to get fetching
collections working.  This has been a "proof of concept" to see what I need.
However, I need to pull that code and rework it to be less taxing on the
database.

And now it turns out that my hack of deleting collections records and adding
them back is failing (Postgresql only) due to this:

#   Failed test 'test_has_many died (DBD::Pg::st execute failed: ERROR:
#   permission denied for relation yello_coll_one
#  [for Statement "DELETE FROM yello_coll_one WHERE yello_id = ?"]

I'll have to look into that next.  Damn.