- Adjusted PG permissions to allow the collection ranking algorithm to work.
The first pass failed because "~ '.+_coll_.+' was matching primary keys.  Once
I realized that, I figured out how to restrict that to tables.

- Added is_purged() to Kinetic.pm

- Renamed _collection_table_fields() to _collection_table_columns()

- Added container introspection to Kinetic::Meta::class

- Reworked loading "has_many" collections into a single SQL fetch per
  collection

It's the "has_many" stuff which once again took the longest time.  I tried to
get this to be loaded "on demand" (when the $object->$collection method is
called), but I wound up backing out my changes to make this happen as I made
some mistakes in figuring out the timing of everything (i.e., when the
attribute information needs to be set).

Still, I have a better idea of how this works and I think I can get it to
work, but I'm concerned that we're doing so much of the querying for folks
that if someone needs a custom query which is really complicated, they might
have problems.  However, if we can push enough business logic to the database,
we have the reflective capability to publish the "official" database API and
potentially allow folks to write custom queries with complicated joins.  This
would require us to have a better way of building objects from those queries.
Right now, _build_object_from_hashref() seems difficult to use.