Collections now allow primitives.
There's a bit of extra work I did to ensure that one can't accidently create a
"bag" instead of a "set" with the collection->from_list method.  It actually
prevented a "bag" earlier, but the error message was delayed an unclear. This
problem still exists when creating a collection from an iterator, but since
iterators are lazy, we won't know about the duplicate value until we pull it
from the iterator.

Renamed $parent_class to $containing_class in attributes.