Laying groundwork for "joins".
- Updated the lexer and parser tests for joins.

The parser tests were failing and I temporarily disabled them while I took a
closer look at the parser.  It shouldn't be too hard to fix.

- Fixed a problem where the parser wasn't generating a valid error message.

- Reworked Kinetic::Store::DB to make query parsing happen earlier.

This broke a lot of tests, but it turns out that these were mostly internal
tests.  Only one actual bug was exposed by this change.

My only concern as I move forward is how to handle the problem with
fully-qualified identifiers.  Right now, they should be properly identified
and handled in the parser, but the logic to do this really belongs in DB.pm
(when we convert the IR to a where clause).  If I push it there, though, I'll
be duplicating functionality.

I think that I may simply add additional data to the search objects and that
*might* solve the problem.