Laying groundwork for "joins".
The first pass involves assembling more metadata for the joins.  Right now,
search objects have been cleaned up tremendously, unused bits have been
removed and tests have been updated.

Now the individual fields in the SQL are fully qualified and this should make
life easier.  

Now when we iterate over the search objects in DB.pm, we can pull off a list
of class keys as needed and that will effectively be the tables we are joining
on.   I think the solution for the WHERE clause at that point will be to
create a new "search_method" which will return an empty array ref for bind
params and the "token" it will return will be something like "person.id =
$other_table.person_id" and then nothing else will hopefully have to change.