Laying groundwork for "joins".
Currently working on ensuring that all column references in our SQL is
"table-qualified" ("person.uuid" instead of "uuid").  This avoids the
inevitable ambiguities which would be introduced when using joins.

Fixed bug in collections where I tried to save the collections before saving
the parent object.  This caused a problem when the parent object was new and
hadn't been saved:  there was no associated ID to put in the collection table.

Currently, there are still two failing tests related to compound searches
generating incorrect SQL for the ORDER BY clause.  It appears that this might
be related to the search_class being set incorrectly by the time we get to the
_contraint_order_by method.  I'll figure this out later, but I'll probably be
taking tomorrow off.  Probably.