Basic joins work.
There was a lot of hacking involved and only one test fails:

    $iterator = $store->query( $class, one => $one );

The problem here crops up and we are searching on an exact contained object
instead of an attribute of the contained object. This fails in
Kinetic::Store::DB::_handle_case_sensitivity.

    if ($search_class) {
        $column = "LOWER($orig_column)"
        if $self->_is_case_sensitive(
            $search_class->attributes($attribute_name)->type
        );
    }

$attribute_name doesn't wind up corresponding with an attribute in the search
class since we wind up getting the wrong search class here.  However, I'm sick
right now and am having a bit of trouble focusing on this last little problem.
Everything else seems to work great.  In fact, the _get_collection method has
pretty much been reduced down to this:

    my $results = $search->package->query(
        "$containing_key.uuid" => $object->uuid,
        { order_by => "$collection_table.seq" }
    );